🚚(lib-video) migrate lti_site/data/queries/useCreateVideo
- migrate lti_site/data/queries/useCreateVideo to lib-video
- Adapt import in LTI project
