+++ src/frontend/apps/lti_site/components/MutationFieldError/index.tsx
@@ -1,26 +1,28 @@
 import { Text } from 'grommet';
+import { IFetchResponseError } from 'lib-components';
 import React from 'react';
 
-interface MutationFieldErrorProps {
-  errors: { [key: string]: string[] | undefined }[];
-  fieldName: string;
+interface MutationFieldErrorProps<T extends unknown> {
+  errors: IFetchResponseError<T>['errors'];
+  fieldName: keyof T;
 }
 
-export const MutationFieldError = ({
+export const MutationFieldError = <T,>({
   errors,
   fieldName,
-}: MutationFieldErrorProps) => (
+}: MutationFieldErrorProps<T>) => (
   <React.Fragment>
-    {errors
-      .filter((errorsObject) => !!errorsObject[fieldName])
-      .map((errorObject, index) => (
-        <React.Fragment key={index}>
-          {errorObject[fieldName]?.map((message) => (
-            <Text color="status-error" key={message}>
-              {message}
-            </Text>
-          ))}
-        </React.Fragment>
-      ))}
+    {errors &&
+      errors
+        .filter((errorsObject) => !!errorsObject[fieldName])
+        .map((errorObject, index) => (
+          <React.Fragment key={index}>
+            {errorObject[fieldName]?.map((message) => (
+              <Text color="status-error" key={message}>
+                {message}
+              </Text>
+            ))}
+          </React.Fragment>
+        ))}
   </React.Fragment>
 );
