+++ src/frontend/packages/lib_video/src/api/index.ts
@@ -13,6 +13,7 @@ export * from './startLive';
 export * from './stopLive';
 export * from './updateLiveParticipants';
 export * from './updateLiveSession';
+export * from './useCreateVideo';
 export * from './useDeleteSharedLiveMedia';
 export * from './useDeleteThumbnail';
 export * from './useDeleteTimedTextTrack';
