+++ src/frontend/packages/lib_video/src/api/useVideoMetadata/index.ts
@@ -1,4 +1,4 @@
-import { metadata } from 'lib-components';
+import { FetchResponseError, metadata } from 'lib-components';
 import { useQuery, UseQueryOptions } from 'react-query';
 
 import { VideoMetadata } from 'types/metadata';
@@ -7,22 +7,23 @@ export const useVideoMetadata = (
   locale: string,
   queryConfig?: UseQueryOptions<
     VideoMetadata,
-    'videos',
+    FetchResponseError<VideoMetadata>,
     VideoMetadata,
     string[]
   >,
 ) => {
   const key = ['videos', locale];
-  return useQuery<VideoMetadata, 'videos', VideoMetadata, string[]>(
-    key,
-    metadata,
-    {
-      refetchInterval: false,
-      refetchIntervalInBackground: false,
-      refetchOnWindowFocus: false,
-      cacheTime: Infinity,
-      staleTime: Infinity,
-      ...queryConfig,
-    },
-  );
+  return useQuery<
+    VideoMetadata,
+    FetchResponseError<VideoMetadata>,
+    VideoMetadata,
+    string[]
+  >(key, metadata, {
+    refetchInterval: false,
+    refetchIntervalInBackground: false,
+    refetchOnWindowFocus: false,
+    cacheTime: Infinity,
+    staleTime: Infinity,
+    ...queryConfig,
+  });
 };
